services/ai/avatar/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pillow>=10.3.0
numpy>=1.26.0
imageio-ffmpeg>=0.5.0
websockets>=12.0
pyyaml>=6.0.1
huggingface_hub>=0.23.0
